requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0,<=2.8.0
torchaudio>=2.0.0,<=2.8.0
torchvision>=0.15.0,<=0.20.0
lightning>=2.0.0

# Audio Processing & Feature Extraction
torchcrepe>=0.3.0
openai-whisper>=20231117
librosa>=0.10.0

# Transformers & Pre-trained Models
transformers>=4.30.0
huggingface-hub>=0.16.0

# Data Processing
pandas>=2.0.0,<=2.3.0
numpy>=1.24.0,<2.0.0

# Configuration & Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Audio/Music Information Retrieval
mir-eval>=0.7

# Scientific Computing (dependencies for audio processing)
scipy>=1.10.0
scikit-learn>=1.3.0
resampy>=0.4.2

# Visualization (optional but useful for training monitoring)
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional utilities
pillow>=10.0.0
requests>=2.31.0

# Optional: for distributed training
# accelerate>=0.20.0